+++ server/api/drinks.js
@@ -19,8 +19,12 @@ router.get('/', async (req, res, next) => {
 
 router.post('/', async (req, res, next) => {
   try {
-    const newDrink = await Drink.create(req.body)
-    res.json(newDrink)
+    if (req.user && req.user.isAdmin) {
+      const newDrink = await Drink.create(req.body)
+      res.json(newDrink)
+    } else {
+      res.redirect('/')
+    }
   } catch (error) {
     next(error)
   }
@@ -28,13 +32,17 @@ router.post('/', async (req, res, next) => {
 
 router.put('/:drinkId', async (req, res, next) => {
   try {
-    const [_, drink] = await Drink.update(req.body, {
-      returning: true,
-      where: {
-        id: req.params.drinkId
-      }
-    })
-    res.send(drink[0].dataValues)
+    if (req.user && req.user.isAdmin) {
+      const [_, drink] = await Drink.update(req.body, {
+        returning: true,
+        where: {
+          id: req.params.drinkId
+        }
+      })
+      res.send(drink[0].dataValues)
+    } else {
+      res.redirect('/')
+    }
   } catch (error) {
     next(error)
   }
@@ -42,12 +50,16 @@ router.put('/:drinkId', async (req, res, next) => {
 
 router.delete('/:drinkId', async (req, res, next) => {
   try {
-    await Drink.destroy({
-      where: {
-        id: req.params.drinkId
-      }
-    })
-    res.status(204).end()
+    if (req.user && req.user.isAdmin) {
+      await Drink.destroy({
+        where: {
+          id: req.params.drinkId
+        }
+      })
+      res.status(204).end()
+    } else {
+      res.redirect('/')
+    }
   } catch (error) {
     next(error)
   }
